requirements.txt
python-dotenv==0.17.1
pyTelegramBotAPI==3.7.9